test: Split monolith test files into per-module test files matching new…
- src/webview/handlers/__tests__/streaming-handlers.test.ts
- src/webview/handlers/__tests__/tool-execution-handlers.test.ts
- src/webview/handlers/__tests__/state-handlers.test.ts
- src/webview/handlers/__tests__/ui-notification-handlers.test.ts
- src/webview/handlers/__tests__/parallel-tools.test.ts
- src/webview/render/__tests__/streaming.test.ts
- src/webview/render/__tests__/html-builders.test.ts
- src/webview/render/__tests__/batches.test.ts

GSD-Task: S02/T04
